Update the Capitomcat version to 1.2.1
Revert back the executing tomcat command in background. gives choice option whether use background or not
It really hard to know which OS is work with pty true/false for tomcat start script. it depends.... please refer to http://capistranorb.com/documentation/faq/why-does-something-work-in-my-ssh-session-but-not-in-capistrano/
Some of the OS may only works with background script. ex) CentOS 5.6